feat: handle input events
This took me a while. Turned out I registered events multiple times.
Lucky me that I checked the number of calls for the handlers to notice
this. In order to properly remove the event listener I have to have
named functions. Therefore A little refactor was involved here as well.

Signed-off-by: André Jaenisch <andre.jaenisch@posteo.de>