feat: render SVG with world
I decided to start with a visual representation and read up about how to
draw isometric. Turns out it means that all cube length have the same
length (duh). I want to display the perspective with 30 degrees and
picked a face to start with. This oen needs to be intersected (because I
want to have a 4x4 grid on each face). But it works as a proof of
concept.

Signed-off-by: André Jaenisch <andre.jaenisch@posteo.de>