build: add esm-seedrandom as dependency
In order to build an AI for bots to act as a player I am going to have
the need of randomness. Now Math.random() isn't seedable, that is, the
tests are likely going to fail _some of the time_.

Therefore I spend a little time looking for proper seedable pseudo
random number generator that are ESM compatible. jsrand is often
suggested but is CommonJS only (or have to be included as a script).

I ended up going with esm-seedrandom.

I also learned about a proposal to have this functionality as part of
the JavaScript language itself, which is amazing!

Signed-off-by: André Jaenisch <andre.jaenisch@posteo.de>