refactor: move check to new reducer
Instead of having a god reducer I'm dispatching a second action after
the first one. This works here because I don't have to compare the old
and the new state in the same reducer (and even if I had to, I could
have passed the old state as payload to the reducer).

Signed-off-by: André Jaenisch <andre.jaenisch@posteo.de>